pyproject.toml
[tool.poetry]
name = "agent-zero"
version = "0.1.0"
description = "General AI Agent"
authors = ["GaryOcean428 <braden.lang77@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ansio = "^0.0.2"
python-dotenv = "^1.0.1"
langchain = "^0.3.1"
langchain-groq = "^0.2.0"
langchain-huggingface = "^0.1.0"
langchain-openai = "^0.2.1"
langchain-community = "^0.3.1"
langchain-anthropic = "^0.2.1"
langchain-google-genai = "^2.0.0"
webcolors = "^24.8.0"
sentence-transformers = "^3.1.1"
docker = "^7.1.0"
paramiko = "^3.5.0"
duckduckgo-search = "^6.2.13"
inputimeout = "^1.0.4"
newspaper3k = "^0.2.8"
beautifulsoup4 = "^4.12.3"
lxml-html-clean = "^0.2.2"
pynput = "^1.7.7"
pypdf = "^5.0.0"
flask = {extras = ["async"], version = "^3.0.3"}
flask-basicauth = "^0.2.0"
faiss-cpu = "^1.8.0.post1"
openai = "^1.48.0"
google-cloud-translate = "^3.16.0"
anthropic = "^0.34.2"
black = "^24.8.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false


[flake8]
max-line-length = 120
